+++ R/utils.R
@@ -7,6 +7,13 @@ os_is_windows <- function() {
   isTRUE(.Platform$OS.type == "windows")
 }
 
+#' Check for macOS
+#' @return `TRUE` if OS is macOS, `FALSE` if not.
+#' @noRd
+os_is_macos <- function() {
+  isTRUE(Sys.info()[["sysname"]] == "Darwin")
+}
+
 #' Famous helper for switching on `NULL` or zero length
 #' @noRd
 #' @param x,y Any \R objects.
@@ -175,10 +182,9 @@ generate_file_names <-
     if (!is.null(ids)) {
       new_names <- paste0(new_names, "-", ids)
     }
-
     if (random) {
-      tf <- tempfile()
-      rand <- substr(tf, nchar(tf) - 5, nchar(tf))
+      rand_num_pid <- as.integer(runif(1, min = 0, max = 1E7)) + Sys.getpid()
+      rand <- format(as.hexmode(rand_num_pid) , width = 6)
       new_names <- paste0(new_names, "-", rand)
     }