+++ tests/testthat/test-fit-shared.R
@@ -248,3 +248,134 @@ test_that("CmdStanArgs erorrs if idx is out of proc_ids range", {
     "Index \\(5\\) exceeds number of CmdStan processes \\(4\\)."
   )
 })
+
+test_that("no output with refresh = 0", {
+  skip_on_cran()
+  mod <- testing_model("logistic")
+  data_list <- testing_data("logistic")
+  output <- utils::capture.output(tmp <- mod$variational(data = data_list))
+  expect_gt(length(output), 1)
+  output <- utils::capture.output(tmp <- mod$optimize(data = data_list))
+  expect_gt(length(output), 1)
+  output <- utils::capture.output(tmp <- mod$sample(data = data_list, chains = 1))
+  expect_gt(length(output), 1)
+
+  output <- utils::capture.output(tmp <- mod$variational(data = data_list, refresh = 0))
+  expect_equal(length(output), 1)
+  output <- utils::capture.output(tmp <- mod$optimize(data = data_list, refresh = 0))
+  expect_equal(length(output), 1)
+  output <- utils::capture.output(tmp <- mod$sample(data = data_list, refresh = 0, chains = 1))
+  expect_equal(length(output), 3)
+})
+
+test_that("sig_figs works with all methods", {
+  skip_on_cran()
+  m <- "data {
+    int<lower=0> N;
+    int<lower=0> K;
+    int<lower=0,upper=1> y[N];
+    matrix[N, K] X;
+  }
+  parameters {
+    real alpha;
+    vector[K] beta;
+  }
+  model {
+    target += normal_lpdf(alpha | 0, 1);
+    target += normal_lpdf(beta | 0, 1);
+    target += bernoulli_logit_glm_lpmf(y | X, alpha, beta);
+  }
+  generated quantities {
+    real p2 = 0.12;
+    real p5 = 0.12345;
+    real p9 = 0.123456789;
+  }"
+  mod <- cmdstan_model(write_stan_file(m))
+  utils::capture.output(
+    sample <- mod$sample(sig_figs = 2, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(sample$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12, 0.12)
+  )
+  utils::capture.output(
+    sample <- mod$sample(sig_figs = 5, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(sample$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12345, 0.12346)
+  )
+  utils::capture.output(
+    sample <- mod$sample(sig_figs = 10, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(sample$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12345, 0.123456789)
+  )
+  utils::capture.output(
+    variational <- mod$variational(sig_figs = 2, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(variational$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12, 0.12)
+  )
+  utils::capture.output(
+    variational <- mod$variational(sig_figs = 5, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(variational$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12345, 0.12346)
+  )
+  utils::capture.output(
+    variational <- mod$variational(sig_figs = 10, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(variational$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12345, 0.123456789)
+  )
+  utils::capture.output(
+    gq <- mod$generate_quantities(fitted_params = sample, sig_figs = 2, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(gq$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12, 0.12)
+  )
+  utils::capture.output(
+    gq <- mod$generate_quantities(fitted_params = sample, sig_figs = 5, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(gq$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12345, 0.12346)
+  )
+  utils::capture.output(
+    gq <- mod$generate_quantities(fitted_params = sample, sig_figs = 10, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(posterior::subset_draws(gq$draws(), variable = c("p2","p5", "p9"), iteration = 1, chain = 1)),
+    c(0.12, 0.12345, 0.123456789)
+  )
+  utils::capture.output(
+    opt <- mod$optimize(sig_figs = 2, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(opt$mle()[c("p2","p5", "p9")]),
+    c(0.12, 0.12, 0.12)
+  )
+  utils::capture.output(
+    opt <- mod$optimize(sig_figs = 5, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(opt$mle()[c("p2","p5", "p9")]),
+    c(0.12, 0.12345, 0.12346)
+  )
+  utils::capture.output(
+    opt <- mod$optimize(sig_figs = 10, refresh = 0, data = testing_data("logistic"))
+  )
+  expect_equal(
+    as.numeric(opt$mle()[c("p2","p5", "p9")]),
+    c(0.12, 0.12345, 0.123456789)
+  )
+})
+
+
+